# docs/swagger/swagger.yaml
definitions:
  dto.AdminCreditAdjustRequest:
    properties:
      amount:
        type: integer
      description:
        type: string
      user_id:
        type: string
    required:
    - amount
    - description
    - user_id
    type: object
...
  dto.PurchaseCreditRequest:
    properties:
      idempotency_key:
        type: string
      package_code:
        type: string
      payment_method:
        type: string
    required:
    - package_code
    - payment_method
    type: object
...
info:
  contact: {}
  description: Payment service API (Gin)
  title: Modami Payment Service
  version: "1.0"
paths:
  /admin/credits/adjust:
    post:
      consumes:
      - application/json
      parameters:
      - description: Admin credit adjust request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AdminCreditAdjustRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Admin adjust user credits
      tags:
      - Admin
  /credits/balance:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: Wrapped dto.CreditBalanceResponse
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get current user's credit balance
      tags:
      - Credits
  /credits/purchase:
    post:
      consumes:
      - application/json
      description: This endpoint is not implemented; use subscriptions flow instead.
      parameters:
      - description: Purchase credit request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PurchaseCreditRequest'
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "501":
          description: Not Implemented
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Purchase credits (deprecated)
      tags:
      - Credits
  /credits/transactions:
    get:
      parameters:
      - default: 20
        description: Page size
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      - description: Filter by transaction type
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Wrapped dto.CreditTransactionListResponse
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List credit transactions for current user
      tags:
      - Credits
  /invoices:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: Wrapped []dto.InvoiceResponse
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List invoices for current user
      tags:
      - Invoices
  /invoices/{id}:
    get:
      parameters:
      - description: Invoice ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Wrapped dto.InvoiceResponse
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get invoice by ID
      tags:
      - Invoices
  /invoices/{id}/download:
    get:
      parameters:
      - description: Invoice ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: invoice text
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Download invoice (plain text)
      tags:
      - Invoices
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"