# module/payment_gateway/gateway_handler/invoice_handler.go
package gateway_handler

import (
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
	"github.com/modami/be-payment-service/module/core/dto"
	"github.com/modami/be-payment-service/module/core/model"
	"github.com/modami/be-payment-service/module/core/usecases"
)

// InvoiceHandler handles invoice endpoints.
type InvoiceHandler struct {
	invoiceUC *usecases.InvoiceUsecase
}

// NewInvoiceHandler creates an InvoiceHandler.
func NewInvoiceHandler(invoiceUC *usecases.InvoiceUsecase) *InvoiceHandler {
	return &InvoiceHandler{invoiceUC: invoiceUC}
}

// ListInvoices handles GET /api/v1/invoices
// @Summary List invoices for current user
// @Tags Invoices
// @Produce json
// @Security BearerAuth
// @Success 200 {object} map[string]interface{} "Wrapped []dto.InvoiceResponse"
// @Failure 401 {object} map[string]interface{}
// @Failure 403 {object} map[string]interface{}
// @Failure 500 {object} map[string]interface{}
// @Router /invoices [get]
func (h *InvoiceHandler) ListInvoices(c *gin.Context) {
	userID := mustUserID(c)
	invoices, err := h.invoiceUC.ListByUser(c.Request.Context(), userID)
	if err != nil {
		respondErr(c, err)
		return
	}
	dtos := make([]*dto.InvoiceResponse, 0, len(invoices))
	for _, inv := range invoices {
		dtos = append(dtos, modelToInvoiceDTO(inv))
	}
	respond(c, dtos)
}

// GetInvoice handles GET /api/v1/invoices/:id
// @Summary Get invoice by ID
// @Tags Invoices
// @Produce json
// @Security BearerAuth
// @Param id path string true "Invoice ID (uuid)"
// @Success 200 {object} map[string]interface{} "Wrapped dto.InvoiceResponse"
// @Failure 400 {object} map[string]interface{}
// @Failure 401 {object} map[string]interface{}
// @Failure 403 {object} map[string]interface{}
// @Failure 500 {object} map[string]interface{}
// @Router /invoices/{id} [get]
func (h *InvoiceHandler) GetInvoice(c *gin.Context) {
	id, err := uuid.Parse(c.Param("id"))
	if err != nil {
		respondBadRequest(c, "invalid id")
		return
	}
	inv, err := h.invoiceUC.GetInvoice(c.Request.Context(), id)
	if err != nil {
		respondErr(c, err)
		return
	}
	respond(c, modelToInvoiceDTO(inv))
}

// DownloadInvoice handles GET /api/v1/invoices/:id/download
// @Summary Download invoice (plain text)
// @Tags Invoices
// @Produce plain
// @Security BearerAuth
// @Param id path string true "Invoice ID (uuid)"
// @Success 200 {string} string "invoice text"
// @Failure 400 {object} map[string]interface{}
// @Failure 401 {object} map[string]interface{}
// @Failure 403 {object} map[string]interface{}
// @Failure 500 {object} map[string]interface{}
// @Router /invoices/{id}/download [get]
func (h *InvoiceHandler) DownloadInvoice(c *gin.Context) {
	id, err := uuid.Parse(c.Param("id"))
	if err != nil {
		respondBadRequest(c, "invalid id")
		return
	}
	inv, err := h.invoiceUC.GetInvoice(c.Request.Context(), id)
	if err != nil {
		respondErr(c, err)
		return
	}
	// Return plain-text invoice (PDF generation would require a separate library).
	c.Header("Content-Disposition", fmt.Sprintf("attachment; filename=invoice-%s.txt", inv.InvoiceNumber))
	c.Header("Content-Type", "text/plain")
	c.String(http.StatusOK,
		"INVOICE\nNumber: %s\nDate: %s\nTotal: %d VND\n",
		inv.InvoiceNumber,
		inv.CreatedAt.Format("2006-01-02"),
		inv.Total,
	)
}

func modelToInvoiceDTO(inv *model.Invoice) *dto.InvoiceResponse {
	return &dto.InvoiceResponse{
		ID:            inv.ID.String(),
		InvoiceNumber: inv.InvoiceNumber,
		UserID:        inv.UserID.String(),
		Subtotal:      inv.Subtotal,
		TaxAmount:     inv.TaxAmount,
		Total:         inv.Total,
		Description:   inv.Description,
		Status:        inv.Status,
		BillingName:   inv.BillingName,
		BillingEmail:  inv.BillingEmail,
		CreatedAt:     inv.CreatedAt,
	}
}